fix(signals): split issue-side slop triage out of the PR-side shim
src/signals/slop.ts's PR-side exports were converted to a re-export
shim over packages/gittensory-engine/src/signals/slop.ts (#5133), but
the file still carried host-only issue-side triage code alongside the
shim, so scripts/check-engine-parity.ts's shim detection (which only
recognizes a file containing NOTHING but the re-export statement)
treated it as a hand-duplicated twin and failed on the now-intentional
divergence. Moves the issue-side code to src/signals/issue-slop.ts so
slop.ts is a pure shim again.
